put everythings to the navBar
