+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Easipe_MobileApplicationDevelopment"
         tools:targetApi="31">
         <activity
-            android:name=".NavigationBar"
+            android:name=".view.navbar.NavigationBar"
             android:exported="true"
             android:theme="@style/Theme.Easipe_MobileApplicationDevelopment" />
         <activity
