+++ app/src/main/java/com/example/easipe_mobileapplicationdevelopment/view/navbar/NavigationBar.java
@@ -1,4 +1,4 @@
-package com.example.easipe_mobileapplicationdevelopment;
+package com.example.easipe_mobileapplicationdevelopment.view.navbar;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.example.easipe_mobileapplicationdevelopment.R;
 
 public class NavigationBar extends AppCompatActivity {
