+++ app/src/main/res/layout/activity_navigation_bar.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavigationBar">
+    tools:context=".view.navbar.NavigationBar">
 
     <FrameLayout
         android:layout_width="match_parent"
